+++ src/onegov.election_day/onegov/election_day/tests/test_views.py
@@ -472,7 +472,7 @@ def test_view_notifications_votes(election_day_app):
     assert "Benachrichtigungen auslösen" not in client.get('/manage/votes')
     assert "Benachrichtigungen auszulösen" not in upload_vote(client, False)
 
-    election_day_app.principal.webhooks = ['http://example.com/1']
+    election_day_app.principal.webhooks = {'http://example.com/1': None}
 
     assert "Benachrichtigungen auslösen" in client.get('/manage/votes')
     assert "Benachrichtigungen auszulösen" in upload_vote(client, False)
@@ -504,7 +504,7 @@ def test_view_notifications_elections(election_day_app_gr):
         client, False
     )
 
-    election_day_app_gr.principal.webhooks = ['http://example.com/1']
+    election_day_app_gr.principal.webhooks = {'http://example.com/1': None}
 
     assert "Benachrichtigungen auslösen" in client.get('/manage/elections')
     assert "Benachrichtigungen auszulösen" in upload_majorz_election(